fix: project section
