+++ src/components/MainPage/Projects/index.tsx
@@ -1,9 +1,13 @@
-
+import ArrowBottomIcon from '@/components/common/ArrowBottomIcon';
+import MagnifierIcon from '@/components/common/MagnifierIcon/index';
 import ProjectCard from '@/components/common/ProjectCard';
 import { projects } from './projects';
 import styles from './styles.module.scss';
-
+function generateRandomNumber() {
+  return Math.floor(Math.random() * 9000) + 1000;
+}
 const Projects = () => {
+  // const project = projects.map((project) => console.log(project));
   return (
     <section className={styles['projects-section']} id="projects">
       <div className={`container`}>
@@ -26,49 +30,21 @@ const Projects = () => {
               type="submit"
               className={styles['projects-section__form__button']}
             >
-              <svg
-                className={styles['projects-section__form__icon']}
-                id="magnifier"
-                fill="none"
-                width="24"
-                height="24"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M11 19a8 8 0 1 0 0-16 8 8 0 0 0 0 16ZM21 21l-4.35-4.35"
-                  stroke="#464646"
-                  strokeWidth="2"
-                  stroke-Linecap="round"
-                  stroke-Linejoin="round"
-                />
-              </svg>
+              <MagnifierIcon />
             </button>
           </form>
         </div>
         <>
           <ul className={styles['projects-section__projects-container']}>
-            <ProjectCard projects={projects} />
+            {projects.map((project) => (
+              <ProjectCard key={generateRandomNumber()} project={project} />
+            ))}
           </ul>
         </>
         <div className={styles['projects-section__load-more__container']}>
           <button className={styles['projects-section__load-more']}>
             Більше проєктів
-            <svg
-              className={styles['header__lang-arrow']}
-              width="16"
-              height="9"
-              viewBox="0 0 16 9"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M15 1L8 8L1 1"
-                stroke="#232323"
-                strokeWidth="1.5"
-                strokeLinecap="round"
-                stroke-Linejoin="round"
-              />
-            </svg>
+            <ArrowBottomIcon />
           </button>
         </div>
       </div>
