+++ src/components/common/ArrowBottomIcon/index.tsx
@@ -1,0 +1,20 @@
+const ArrowBottomIcon = () => {
+  return (
+    <svg
+      width="16"
+      height="9"
+      viewBox="0 0 16 9"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M15 1L8 8L1 1"
+        stroke="#232323"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+export default ArrowBottomIcon;
