+++ src/components/common/CloseIcon/index.tsx
@@ -1,0 +1,20 @@
+const CloseIcon = () => {
+  return (
+    <svg
+      id="close"
+      width="24"
+      height="24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M18 6 6 18M6 6l12 12"
+        stroke="#FCFCFC"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+export default CloseIcon;
