+++ src/components/common/LevelIcon/index.tsx
@@ -1,0 +1,20 @@
+const LevelIcon = () => {
+  return (
+    <svg
+      id="level-icon"
+      width="24"
+      height="24"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M18 20V10M12 20V4M6 20v-6"
+        stroke="#FCFCFC"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+export default LevelIcon;
