+++ src/components/common/MagnifierIcon/index.tsx
@@ -1,0 +1,20 @@
+const MagnifierIcon = () => {
+  return (
+    <svg
+      id="magnifier"
+      fill="none"
+      width="24"
+      height="24"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M11 19a8 8 0 1 0 0-16 8 8 0 0 0 0 16ZM21 21l-4.35-4.35"
+        stroke="#464646"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
+export default MagnifierIcon;
