+++ src/components/common/ProjectCard/styles.module.scss
@@ -1,4 +1,3 @@
-@import '../../../styles/abstracts/_variables.scss';
 .projects-section {
   &__projects-item {
     width: 37.8rem;
