+++ src/utils/formatDate.ts
@@ -1,0 +1,23 @@
+const formatDate = (dateString: string) => {
+  const monthNames = [
+    'січня',
+    'лютого',
+    'березня',
+    'квітня',
+    'травня',
+    'червня',
+    'липня',
+    'серпня',
+    'вересня',
+    'жовтня',
+    'листопада',
+    'грудня',
+  ];
+  const date = new Date(dateString);
+  const day = date.getDate();
+  const month = monthNames[date.getMonth()];
+  const year = date.getFullYear();
+
+  return `${day} ${month} ${year}`;
+};
+export default formatDate;
